Major feature batch: vehicle UX, notifications, profile, images, fixes
Vehicle details:
- Replaced status spinner with contextual UI: availability toggle
  (Disponible/Indisponible), assign driver spinner, current driver card
  with Retirer button, En trajet lock banner
- Fixed driver spinner always being enabled regardless of status
- Depot map pin now 2x larger so it's visible under vehicle pins

Trip flow:
- startTrip() now takes onStarted callback so trip button flips
  immediately without waiting for the next fragment resume
- Vehicle status resets to Assigné after trip ends

Notifications:
- Replaced PeriodicWorkRequest with self-rescheduling OneTimeWorkRequest
  to bypass 15-minute minimum (now runs every 3 minutes)
- Admin notified for all entretiens; chauffeur only for their vehicle

Chauffeur/vehicle lists:
- vehiculeAffecte resolved client-side in ChauffeursFragment
- Vehicle list status chips use correct statuses (Assigné, En trajet)
- Chauffeur dropdown filters to unassigned drivers only

Profile page (admin + chauffeur):
- Replaced nom complet + email fields with Prénom, Nom, Identifiant,
  Mot de passe — loads from and saves to backend
- New GET/PUT /api/admins/{id} endpoint

Seed data:
- Fixed trajet INSERT SQL (invalid row-constructor syntax broke all 8 rows)
- Real vehicle photos (Renault Master, Dacia Dokker, Peugeot Boxer,
  Iveco Daily, Ford Transit) embedded as Base64 in data.sql
- Real accident/damage photos for the 4 seeded incidents
- New vehicle statuses (Assigné replaces En mission)
- Notification test entretiens (e_t1, e_t2)

Admin home:
- Replaced hardcoded "31 Mai 2026" with live today's date

+++ app/src/main/java/com/fleettracking/app/model/Trajet.java
@@ -10,7 +10,7 @@ public class Trajet {
     public String arrivee;
     public String heureDepart;
     public String heureArrivee;
-    public int distanceKm;
+    public double distanceKm;
     public String duree;
     public int vitesseMoyenne;
     public double consommation;   // L/100km
